requirements.txt
asgiref==3.2.10
Django==2.2
django-markdownx==3.0.1
Markdown==3.3.2
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
